server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
Flask-Marshmallow
marshmallow-sqlalchemy
python-dotenv
psycopg2-binary
passlib[bcrypt]
Flask-Cors
boto3
redis
requests
Flask-APScheduler